% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_details.R
\name{get_details}
\alias{get_details}
\title{Retrieve observations from series}
\usage{
get_details(name = NULL, group = F)
}
\arguments{
\item{name}{A character of length 1.}

\item{group}{A Boolean.}
}
\value{
A list.
}
\description{
Retrieve observations from series
}
\details{
\emph{name} refers to the series or series group name identifiable on the Bank of Canada's website or via \code{get_list}.
}
