% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_list.R
\name{get_list}
\alias{get_list}
\title{List possible series or groups.}
\usage{
get_list(type = c("series", "groups"))
}
\arguments{
\item{type}{Either \code{series} or \code{groups}.}
}
\value{
A \code{tibble}.
}
\description{
List possible series or groups.
}
