% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group.R
\name{get_group}
\alias{get_group}
\title{Retrieve observations from series}
\usage{
get_group(name = NULL)
}
\arguments{
\item{name}{A character of length 1.}
}
\value{
A \code{tibble}.
}
\description{
Retrieve observations from series
}
