% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_series.R
\name{get_series}
\alias{get_series}
\title{Retrieve observations from series}
\usage{
get_series(name = NULL, ...)
}
\arguments{
\item{name}{A character of at least length 1.}
}
\value{
A \code{tibble}.
}
\description{
Retrieve observations from series
}
