% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valet.R
\name{valet}
\alias{valet}
\title{Retrieve Valet response}
\usage{
valet(name = NULL, group = F, ...)
}
\arguments{
\item{name}{A character of at least length 1 indicating the series or series
group to retrieve.}

\item{group}{Boolean indicating whether the \code{name} is a series or a
series group.}

\item{...}{Query parameters from other methods.}
}
\value{
A \code{valet} object.
}
\description{
\code{valet} is the core back-end to get responses from the Bank of Canada
API.
}
\examples{
valet(name = "GBPP", group = T, recent_weeks = 2)

}
