% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valet.R
\name{valet}
\alias{valet}
\title{Retrieve Valet response.}
\usage{
valet(name = NULL, group = F, ...)
}
\arguments{
\item{name}{A character of at least length 1.}

\item{group}{Boolean.}
}
\value{
A \code{valet} object.
}
\description{
Retrieve Valet response.
}
